+++ src/main/java/com/example/SWP_Project_BackEnd/Repository/KoiFarmRepository.java
@@ -9,6 +9,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface KoiFarmRepository extends JpaRepository<KoiFarm, Long> {
+public interface KoiFarmRepository extends JpaRepository<KoiFarm, Integer> {
 }
